requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-multipart==0.0.6
google-generativeai==0.3.2
PyPDF2==3.0.1
pydantic==2.6.1
langgraph==0.0.13
pytest==7.4.4
pytest-asyncio==0.23.5
httpx==0.26.0